Refactor superagent transport: extract bind-transport method
